Temporarily remove native ble virtual interface
