+++ transport_plugins/native_ble/RELEASE.md
@@ -2,6 +2,10 @@
 
 All major changes in each released version of the native BLE transport plugin are listed here.
 
+## HEAD
+
+- Temporarily remove `virtual_ble` interface as it is ported to be a DeviceServer.
+
 ## 2.0.3
 
 - Implement proper dependency major version limits.
