+++ transport_plugins/native_ble/test/test_virtual_interface.py
@@ -11,6 +11,49 @@ from iotile.core.hw.reports import IOTileReportParser
 from test.devices_factory import build_report_device, build_tracing_device, get_report_device_string, \
     get_tracing_device_string
 
+pytestmark = pytest.mark.skip
+
+# from iotile_transport_native_ble.virtual_ble import NativeBLEVirtualInterface
+
+
+@pytest.fixture(scope="function")
+def virtual_interface(mock_bable, request):
+    """Initialize the NativeBLEVirtualInterface with a given device."""
+    port = request.param['port']
+    config = {
+        'port': port
+    }
+
+    device = request.param['device']
+
+    interface = NativeBLEVirtualInterface(config)
+    interface.start(device)
+
+    # Call the process function periodically (in a separated thread)
+    def _call_process():
+        while True:
+            interface.process()
+            time.sleep(0.1)
+
+    process_thread = threading.Thread(target=_call_process, name='VirtualInterfaceProcessThread')
+    process_thread.daemon = True
+    process_thread.start()
+
+    return mock_bable, interface
+
+
+@pytest.fixture(scope="function")
+def connected_virtual_interface(virtual_interface):
+    """Create an already connected NativeBLEVirtualInterface."""
+    mock_bable, interface = virtual_interface
+    device_address = '11:11:11:11:11:11'
+
+    mock_bable.simulate_connected_event(interface.controller_id, device_address)
+    assert interface.connected is True
+
+    return mock_bable, interface
+
+
 
 @pytest.mark.parametrize('virtual_interface', [{'device': build_report_device(), 'port': None}], indirect=True)
 def test_start_stop(virtual_interface):
